feat(frontend):New style file
